Extract collection-family test helpers into CollectionTestUtils
