+++ libraries/tools/kotlin-stdlib-gen/src/generators/test/IsSortedSampleGenerator.kt
@@ -11,7 +11,7 @@ import java.io.BufferedWriter
 
 object IsSortedSampleGenerator {
     fun generate() {
-        forEachIsSortedFamily { family, primitive -> generate(family, primitive) }
+        forEachCollectionFamily { family, primitive -> generate(family, primitive) }
     }
 
     private fun generate(family: Family, primitive: PrimitiveType? = null) {
